Extract status assertions to trait
